refactoring solvers, testing improvements to pdas, testing admm solver.

+++ src/proxtf/wpdas.h
@@ -6,19 +6,19 @@
 //#endif
 
 /* main routine for pdas l1tf solver */
-int weighted_pdas(const int n,
-	          const double *y,
-	          const double *wi,	       
-	          const double lambda,
-	          double *x,
-                  double *z,
-                  int *iter,
-                  double p,
-                  const int m,
-                  const double delta_s,
-                  const double delta_e,
-                  const int maxiter,
-                  const int verbose);
+short weighted_pdas(const int n,
+                    const double *y,
+                    const double *wi,	       
+                    const double lambda,
+                    double *x,
+                    double *z,
+                    int *iter,
+                    double p,
+                    const int m,
+                    const double delta_s,
+                    const double delta_e,
+                    const int maxiter,
+                    const int verbose);
 
 //#ifdef  __cplusplus
 //}
